requirements.txt
altair==6.0.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
gitdb==4.0.12
gitpython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
httplib2==0.31.2
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
markupsafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.18.0
nest-asyncio==1.6.0
numpy==2.4.3
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prompt-toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
traitlets==5.14.3
typing-extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
altair==6.0.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
gitdb==4.0.12
gitpython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
httplib2==0.31.2
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
markupsafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.18.0
nest-asyncio==1.6.0
numpy==2.4.3
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
playwright==1.58.0
prompt-toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydeck==0.9.1
pyee==13.0.1
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
traitlets==5.14.3
typing-extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
